@@ -1,17 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Net.Http;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
+using Utf8Json;
 
 class Program
 {
@@ -37,7 +31,7 @@ class Program
     }
 }
 
-class Response
+public struct Response
 {
     public string Id { get; set; }
     public string Name { get; set; }
@@ -60,22 +54,16 @@ class Startup
     {
         app.Run(async ctx =>
         {
-            using (var rsp = await _http.GetAsync("/data"))
-            {
-                var str = await rsp.Content.ReadAsStringAsync();
+            await using var rsp = await _http.GetStreamAsync("/data");
 
-                // deserialize
-                var obj = JsonConvert.DeserializeObject<Response>(str);
+            // deserialize
+            var obj = await JsonSerializer.DeserializeAsync<Response>(rsp);
 
-                // serialize
-                var json = JsonConvert.SerializeObject(obj);
-
-                ctx.Response.ContentType = "application/json";
-                await ctx.Response.WriteAsync(json);
-            }
+            // serialize
+            ctx.Response.ContentType = "application/json";
+            await JsonSerializer.SerializeAsync(ctx.Response.Body, obj);
         });
     }
-
     public void Configure(IApplicationBuilder app)
     {
         app.Map("/test", HandleTest);
